% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BobRoss-package.R
\docType{package}
\name{BobRoss-package}
\alias{BobRoss}
\alias{BobRoss-package}
\title{BobRoss: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aditya Santosh Rao \email{asan0096@student.monash.edu}

}
\keyword{internal}
