% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BobRoss-package.R
\docType{package}
\name{BobRoss}
\alias{BobRoss-package}
\alias{BobRoss}
\title{BobRoss: Analyze Bob Ross's Paintings with a Shiny App}
\description{
The \strong{BobRoss} package provides an interactive Shiny app to explore and analyze the colors used in Bob Ross's \emph{The Joy of Painting} series.
It includes detailed color information for each painting and provides tools to visualize color trends over time. The package also provides access
to a dataset containing metadata on the paintings, such as the season, episode, and color details.
}
\section{Key Functions}{

\itemize{
\item \code{\link{loadbob}}: Launches the Shiny app for exploring Bob Ross's painting colors.
\item \code{\link{bob_ross}}: Accesses the dataset containing information about the paintings and their colors.
}
}

\section{Dataset Overview}{

The dataset includes information on over 400 paintings, covering variables such as:
\itemize{
\item \strong{img_src}: URL to the image of the painting.
\item \strong{painting_title}: Title of the painting.
\item \strong{season}: Season number.
\item \strong{episode}: Episode number within the season.
\item \strong{num_colors}: Number of colors used in the painting.
\item \strong{youtube_src}: URL to the YouTube video of the episode.
\item \strong{colors}: List of colors used in the painting.
\item \strong{color_hex}: Hexadecimal values for the colors used.
}
}

\section{How to Use the Package}{

\enumerate{
\item Install the package:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github("ETC5523-2024/assignment-4-packages-and-shiny-apps-Raodityaaa/BobRoss")
}\if{html}{\out{</div>}}
\enumerate{
\item Load the package:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(BobRoss)
}\if{html}{\out{</div>}}
\enumerate{
\item Launch the Shiny app to explore Bob Ross's paintings:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{loadbob()
}\if{html}{\out{</div>}}
\enumerate{
\item Access the painting dataset for custom analysis:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data <- bob_ross()
head(data)
}\if{html}{\out{</div>}}
}

\section{Author}{

Aditya Rao
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2024/assignment-4-packages-and-shiny-apps-Raodityaaa/}
  \item \url{https://etc5523-2024.github.io/assignment-4-packages-and-shiny-apps-Raodityaaa/}
}

}
\author{
\strong{Maintainer}: Aditya Santosh Rao \email{asan0096@student.monash.edu}

}
