% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{loadbob}
\alias{loadbob}
\title{Loading the Rshiny app}
\usage{
loadbob()
}
\description{
This function launches the RShiny app which allows the user to explore the "Joy of Painting: Bob Ross'
data for analysis and research purposes.
}
