% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BobRoss-package.R
\docType{package}
\name{BobRoss-package}
\alias{BobRoss}
\alias{BobRoss-package}
\title{BobRoss: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2024/assignment-4-packages-and-shiny-apps-Raodityaaa/}
  \item \url{https://etc5523-2024.github.io/assignment-4-packages-and-shiny-apps-Raodityaaa/}
}

}
\author{
\strong{Maintainer}: Aditya Santosh Rao \email{asan0096@student.monash.edu}

}
\keyword{internal}
